Split accessor tests
